+++ src/datasource/DronerDatasource.ts
@@ -59,7 +59,6 @@ export class DronerDatasource {
       delete data.dronerArea["subdistrictId"];
       delete data.dronerArea["distance"];
     }
-    console.log("data", JSON.stringify(data));
     return httpClient
       .patch(BASE_URL + "/droner/" + data.id, data)
       .then((response) => {